+++ src/trx/game/objects/creatures/natla.c
@@ -140,7 +140,10 @@ static void M_Control(const int16_t item_num)
                 natla->flags = 0;
                 timer = 0;
                 item->hit_points = M_GetStage2HitPoints(item);
-                Music_Play(MX_NATLA_SPEECH, MPM_NO_REPEAT);
+                const MUSIC_PLAY_MODE mode =
+                    g_Config.audio.fix_speeches_killing_music ? MPM_OVERLAY
+                                                              : MPM_NO_REPEAT;
+                Music_Play(MX_NATLA_SPEECH, mode);
             } else {
                 if (g_Config.gameplay.target_mode == TLM_SEMI
                     || g_Config.gameplay.target_mode == TLM_NONE) {